requirements.txt
pathlib==1.0.1
mysql-connector-python==8.0.32
pyfiglet==0.8.post1
numpy==1.24.1
pandas==1.5.2